Places: Support sub-km distances when searching for locations #3558

+++ internal/form/search_photos_geo.go
@@ -46,7 +46,7 @@ type SearchPhotosGeo struct {
 	Olc       string    `form:"olc" example:"olc:8FWCHX7W+" notes:"OLC Position (Open Location Code)"`
 	Lat       float64   `form:"lat" example:"lat:41.894043" notes:"GPS Position (Latitude)"`
 	Lng       float64   `form:"lng" example:"lng:-87.62448" notes:"GPS Position (Longitude)"`
-	Dist      uint      `form:"dist" example:"dist:50" notes:"Distance to Position (km)"`
+	Dist      float64   `form:"dist" example:"dist:50" notes:"Distance to Position (km)"`
 	Latlng    string    `form:"latlng" notes:"GPS Bounding Box (Lat N, Lng E, Lat S, Lng W)"`
 	Person    string    `form:"person"`   // Alias for Subject
 	Subjects  string    `form:"subjects"` // Text
